refactor: #13-구조 변경 반영
